Fixed major typo and added log odds to computations
When sampling the posterior for each v_alpha (the mixing variable for Laplace/t distributions), the code was sampling the subsets gamma_alpha = 0 and gamma_alpha = 1 separately..
However, the code was using the entirety of alpha (both active and inactive) in sampling each part.
Should now be fixed to correctly use alpha[gamma_alpha==0] and alpha[gamma_alpha==1], and similarly for beta.

I am not sure what impact this had on the algorithm. It has none on normal priors. I only noticed this problem because the code broke down for very high correlation (0.95), and I tracked this to NAs appearing because of this problem, but I don't fully understand why this would happen only in the high correlation case.

Before finding the real problem, I thought the problem was numerical instability and changed the gamma posterior probability calculations to use log odds (as it involved using dnorm, which for extreme values could return 0), though it turns out that wasn't the problem.
In any case, the code uses log odds for calculation of posterior probabilities now, as it is less likely to have issues with numerical instability (in particular for coefficients much larger than the slab variance).