small changes
Changed param>0 to sum(param)>0 to get rid of the warning

param=0 on the initial iteration to change the sampling distribution to the prior for v_alpha, in other iterations param is a vector with all positive values. Previously it checked the first value > 0 and continued, but gave a warning each time.